chore(restructure): split main.py -> video/* + ml/* + ingest/youtube.py
Phase 1 step 7 (biggest split): main.py is broken into eight modules:
- openshorts/video/tracking.py        SmoothedCameraman, SpeakerTracker
- openshorts/video/scene_analysis.py  detect_scenes, analyze_scenes_strategy
- openshorts/video/reframing.py       create_general_frame
- openshorts/video/pipeline.py        process_video_to_vertical (the hot loop)
- openshorts/ml/detection.py          detect_face_candidates, detect_person_yolo
- openshorts/ml/transcription.py      transcribe_video
- openshorts/ml/viral_extraction.py   GEMINI_PROMPT_TEMPLATE, get_viral_clips
- openshorts/ingest/youtube.py        download_youtube_video, sanitize_filename

main.py becomes a thin shim that re-exports the public surface for backwards
compatibility AND preserves the CLI entrypoint (`python main.py -i ... -o ...`)
in a private `_cli()` function. Tests stay 62/62 green.
