+++ openshorts/ml/detection.py
@@ -1,0 +1,73 @@
+"""Face and person detection: MediaPipe BlazeFace (primary) + YOLOv8 (fallback)."""
+
+import cv2
+from ultralytics import YOLO
+import mediapipe as mp
+
+# Load the YOLO model once (Keep for backup or scene analysis if needed)
+model = YOLO('yolov8n.pt')
+
+# --- MediaPipe Setup ---
+# Use standard Face Detection (BlazeFace) for speed
+mp_face_detection = mp.solutions.face_detection
+face_detection = mp_face_detection.FaceDetection(model_selection=1, min_detection_confidence=0.5)
+
+
+def detect_face_candidates(frame):
+    """
+    Returns list of all detected faces using lightweight FaceDetection.
+    """
+    height, width, _ = frame.shape
+    rgb_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+    results = face_detection.process(rgb_frame)
+
+    candidates = []
+
+    if not results.detections:
+        return []
+
+    for detection in results.detections:
+        bboxC = detection.location_data.relative_bounding_box
+        x = int(bboxC.xmin * width)
+        y = int(bboxC.ymin * height)
+        w = int(bboxC.width * width)
+        h = int(bboxC.height * height)
+
+        candidates.append({
+            'box': [x, y, w, h],
+            'score': w * h # Area as score
+        })
+
+    return candidates
+
+
+def detect_person_yolo(frame):
+    """
+    Fallback: Detect largest person using YOLO when face detection fails.
+    Returns [x, y, w, h] of the person's 'upper body' approximation.
+    """
+    # Use the globally loaded model
+    results = model(frame, verbose=False, classes=[0]) # class 0 is person
+
+    if not results:
+        return None
+
+    best_box = None
+    max_area = 0
+
+    for result in results:
+        boxes = result.boxes
+        for box in boxes:
+            x1, y1, x2, y2 = [int(i) for i in box.xyxy[0]]
+            w = x2 - x1
+            h = y2 - y1
+            area = w * h
+
+            if area > max_area:
+                max_area = area
+                # Focus on the top 40% of the person (head/chest) for framing
+                # This approximates where the face is if we can't detect it directly
+                face_h = int(h * 0.4)
+                best_box = [x1, y1, w, face_h]
+
+    return best_box
