+++ openshorts/ml/viral_extraction.py
@@ -1,0 +1,137 @@
+"""Gemini 2.5 Flash viral-moment extraction: picks 3-15 short clips from a transcript."""
+
+import os
+import json
+from google import genai
+
+
+GEMINI_PROMPT_TEMPLATE = """
+You are a senior short-form video editor. Read the ENTIRE transcript and word-level timestamps to choose the 3–15 MOST VIRAL moments for TikTok/IG Reels/YouTube Shorts. Each clip must be between 15 and 60 seconds long.
+
+⚠️ FFMPEG TIME CONTRACT — STRICT REQUIREMENTS:
+- Return timestamps in ABSOLUTE SECONDS from the start of the video (usable in: ffmpeg -ss <start> -to <end> -i <input> ...).
+- Only NUMBERS with decimal point, up to 3 decimals (examples: 0, 1.250, 17.350).
+- Ensure 0 ≤ start < end ≤ VIDEO_DURATION_SECONDS.
+- Each clip between 15 and 60 s (inclusive).
+- Prefer starting 0.2–0.4 s BEFORE the hook and ending 0.2–0.4 s AFTER the payoff.
+- Use silence moments for natural cuts; never cut in the middle of a word or phrase.
+- STRICTLY FORBIDDEN to use time formats other than absolute seconds.
+
+VIDEO_DURATION_SECONDS: {video_duration}
+
+TRANSCRIPT_TEXT (raw):
+{transcript_text}
+
+WORDS_JSON (array of {{w, s, e}} where s/e are seconds):
+{words_json}
+
+STRICT EXCLUSIONS:
+- No generic intros/outros or purely sponsorship segments unless they contain the hook.
+- No clips < 15 s or > 60 s.
+
+OUTPUT — RETURN ONLY VALID JSON (no markdown, no comments). Order clips by predicted performance (best to worst). In the descriptions, ALWAYS include a CTA like "Follow me and comment X and I'll send you the workflow" (especially if discussing an n8n workflow):
+{{
+  "shorts": [
+    {{
+      "start": <number in seconds, e.g., 12.340>,
+      "end": <number in seconds, e.g., 37.900>,
+      "video_description_for_tiktok": "<description for TikTok oriented to get views>",
+      "video_description_for_instagram": "<description for Instagram oriented to get views>",
+      "video_title_for_youtube_short": "<title for YouTube Short oriented to get views 100 chars max>",
+      "viral_hook_text": "<SHORT punchy text overlay (max 10 words). MUST BE IN THE SAME LANGUAGE AS THE VIDEO TRANSCRIPT. Examples: 'POV: You realized...', 'Did you know?', 'Stop doing this!'>"
+    }}
+  ]
+}}
+"""
+
+
+def get_viral_clips(transcript_result, video_duration):
+    print("🤖  Analyzing with Gemini...")
+
+    api_key = os.getenv("GEMINI_API_KEY")
+    if not api_key:
+        print("❌ Error: GEMINI_API_KEY not found in environment variables.")
+        return None
+
+    client = genai.Client(api_key=api_key)
+
+    # We use gemini-2.5-flash as requested.
+    model_name = 'gemini-2.5-flash'
+
+    print(f"🤖  Initializing Gemini with model: {model_name}")
+
+    # Extract words
+    words = []
+    for segment in transcript_result['segments']:
+        for word in segment.get('words', []):
+            words.append({
+                'w': word['word'],
+                's': word['start'],
+                'e': word['end']
+            })
+
+    prompt = GEMINI_PROMPT_TEMPLATE.format(
+        video_duration=video_duration,
+        transcript_text=json.dumps(transcript_result['text']),
+        words_json=json.dumps(words)
+    )
+
+    try:
+        response = client.models.generate_content(
+            model=model_name,
+            contents=prompt
+        )
+
+        # --- Cost Calculation ---
+        try:
+            usage = response.usage_metadata
+            if usage:
+                # Gemini 2.5 Flash Pricing (Dec 2025)
+                # Input: $0.10 per 1M tokens
+                # Output: $0.40 per 1M tokens
+
+                input_price_per_million = 0.10
+                output_price_per_million = 0.40
+
+                prompt_tokens = usage.prompt_token_count
+                output_tokens = usage.candidates_token_count
+
+                input_cost = (prompt_tokens / 1_000_000) * input_price_per_million
+                output_cost = (output_tokens / 1_000_000) * output_price_per_million
+                total_cost = input_cost + output_cost
+
+                cost_analysis = {
+                    "input_tokens": prompt_tokens,
+                    "output_tokens": output_tokens,
+                    "input_cost": input_cost,
+                    "output_cost": output_cost,
+                    "total_cost": total_cost,
+                    "model": model_name
+                }
+
+                print(f"💰 Token Usage ({model_name}):")
+                print(f"   - Input Tokens: {prompt_tokens} (${input_cost:.6f})")
+                print(f"   - Output Tokens: {output_tokens} (${output_cost:.6f})")
+                print(f"   - Total Estimated Cost: ${total_cost:.6f}")
+
+        except Exception as e:
+            print(f"⚠️ Could not calculate cost: {e}")
+            cost_analysis = None
+        # ------------------------
+
+        # Clean response if it contains markdown code blocks
+        text = response.text
+        if text.startswith("```json"):
+            text = text[7:]
+        if text.endswith("```"):
+            text = text[:-3]
+        text = text.strip()
+
+        result_json = json.loads(text)
+        if cost_analysis:
+            result_json['cost_analysis'] = cost_analysis
+
+        return result_json
+    except Exception as e:
+        print(f"❌ Gemini Error: {e}")
+        return None
