+++ openshorts/video/pipeline.py
@@ -1,0 +1,195 @@
+"""process_video_to_vertical orchestrator: scenes -> strategy -> per-frame crop -> mux.
+
+The hot loop. Calls scene_analysis, ml.detection, tracking, and reframing per frame,
+then writes raw BGR24 frames to an FFmpeg encoder via stdin and finally muxes audio.
+"""
+
+import os
+import sys
+import time
+import subprocess
+
+import cv2
+from tqdm import tqdm
+
+from openshorts.video.tracking import SmoothedCameraman, SpeakerTracker, ASPECT_RATIO
+from openshorts.video.scene_analysis import (
+    detect_scenes,
+    get_video_resolution,
+    analyze_scenes_strategy,
+)
+from openshorts.video.reframing import create_general_frame
+from openshorts.ml.detection import detect_face_candidates, detect_person_yolo
+
+
+def process_video_to_vertical(input_video, final_output_video):
+    """
+    Core logic to convert horizontal video to vertical using scene detection and Active Speaker Tracking (MediaPipe).
+    """
+    script_start_time = time.time()
+
+    # Define temporary file paths based on the output name
+    base_name = os.path.splitext(final_output_video)[0]
+    temp_video_output = f"{base_name}_temp_video.mp4"
+    temp_audio_output = f"{base_name}_temp_audio.aac"
+
+    # Clean up previous temp files if they exist
+    if os.path.exists(temp_video_output): os.remove(temp_video_output)
+    if os.path.exists(temp_audio_output): os.remove(temp_audio_output)
+    if os.path.exists(final_output_video): os.remove(final_output_video)
+
+    print(f"🎬 Processing clip: {input_video}")
+    print("   Step 1: Detecting scenes...")
+    scenes, fps = detect_scenes(input_video)
+
+    if not scenes:
+        print("   ❌ No scenes were detected. Using full video as one scene.")
+        # If scene detection fails or finds nothing, treat whole video as one scene
+        cap = cv2.VideoCapture(input_video)
+        total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+        cap.release()
+        from scenedetect import FrameTimecode
+        scenes = [(FrameTimecode(0, fps), FrameTimecode(total_frames, fps))]
+
+    print(f"   ✅ Found {len(scenes)} scenes.")
+
+    print("\n   🧠 Step 2: Preparing Active Tracking...")
+    original_width, original_height = get_video_resolution(input_video)
+
+    OUTPUT_HEIGHT = original_height
+    OUTPUT_WIDTH = int(OUTPUT_HEIGHT * ASPECT_RATIO)
+    if OUTPUT_WIDTH % 2 != 0:
+        OUTPUT_WIDTH += 1
+
+    # Initialize Cameraman
+    cameraman = SmoothedCameraman(OUTPUT_WIDTH, OUTPUT_HEIGHT, original_width, original_height)
+
+    # --- New Strategy: Per-Scene Analysis ---
+    print("\n   🤖 Step 3: Analyzing Scenes for Strategy (Single vs Group)...")
+    scene_strategies = analyze_scenes_strategy(input_video, scenes)
+    # scene_strategies is a list of 'TRACK' or 'General' corresponding to scenes
+
+    print("\n   ✂️ Step 4: Processing video frames...")
+
+    command = [
+        'ffmpeg', '-y', '-f', 'rawvideo', '-vcodec', 'rawvideo',
+        '-s', f'{OUTPUT_WIDTH}x{OUTPUT_HEIGHT}', '-pix_fmt', 'bgr24',
+        '-r', str(fps), '-i', '-', '-c:v', 'libx264',
+        '-preset', 'fast', '-crf', '23', '-an', temp_video_output
+    ]
+
+    ffmpeg_process = subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.DEVNULL, stderr=subprocess.PIPE)
+
+    cap = cv2.VideoCapture(input_video)
+    total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+
+    frame_number = 0
+    current_scene_index = 0
+
+    # Pre-calculate scene boundaries
+    scene_boundaries = []
+    for s_start, s_end in scenes:
+        scene_boundaries.append((s_start.get_frames(), s_end.get_frames()))
+
+    # Global tracker for single-person shots
+    speaker_tracker = SpeakerTracker(cooldown_frames=30)
+
+    with tqdm(total=total_frames, desc="   Processing", file=sys.stdout) as pbar:
+        while cap.isOpened():
+            ret, frame = cap.read()
+            if not ret:
+                break
+
+            # Update Scene Index
+            if current_scene_index < len(scene_boundaries):
+                start_f, end_f = scene_boundaries[current_scene_index]
+                if frame_number >= end_f and current_scene_index < len(scene_boundaries) - 1:
+                    current_scene_index += 1
+
+            # Determine Strategy for current frame based on scene
+            current_strategy = scene_strategies[current_scene_index] if current_scene_index < len(scene_strategies) else 'TRACK'
+
+            # Apply Strategy
+            if current_strategy == 'GENERAL':
+                # "Plano General" -> Blur Background + Fit Width
+                output_frame = create_general_frame(frame, OUTPUT_WIDTH, OUTPUT_HEIGHT)
+
+                # Reset cameraman/tracker so they don't drift while inactive
+                cameraman.current_center_x = original_width / 2
+                cameraman.target_center_x = original_width / 2
+
+            else:
+                # "Single Speaker" -> Track & Crop
+
+                # Detect every 2nd frame for performance
+                if frame_number % 2 == 0:
+                    candidates = detect_face_candidates(frame)
+                    target_box = speaker_tracker.get_target(candidates, frame_number, original_width)
+                    if target_box:
+                        cameraman.update_target(target_box)
+                    else:
+                        person_box = detect_person_yolo(frame)
+                        if person_box:
+                            cameraman.update_target(person_box)
+
+                # Snap camera on scene change to avoid panning from previous scene position
+                is_scene_start = (frame_number == scene_boundaries[current_scene_index][0])
+
+                x1, y1, x2, y2 = cameraman.get_crop_box(force_snap=is_scene_start)
+
+                # Crop
+                if y2 > y1 and x2 > x1:
+                    cropped = frame[y1:y2, x1:x2]
+                    output_frame = cv2.resize(cropped, (OUTPUT_WIDTH, OUTPUT_HEIGHT))
+                else:
+                    output_frame = cv2.resize(frame, (OUTPUT_WIDTH, OUTPUT_HEIGHT))
+
+            ffmpeg_process.stdin.write(output_frame.tobytes())
+            frame_number += 1
+            pbar.update(1)
+
+    ffmpeg_process.stdin.close()
+    stderr_output = ffmpeg_process.stderr.read().decode()
+    ffmpeg_process.wait()
+    cap.release()
+
+    if ffmpeg_process.returncode != 0:
+        print("\n   ❌ FFmpeg frame processing failed.")
+        print("   Stderr:", stderr_output)
+        return False
+
+    print("\n   🔊 Step 5: Extracting audio...")
+    audio_extract_command = [
+        'ffmpeg', '-y', '-i', input_video, '-vn', '-acodec', 'copy', temp_audio_output
+    ]
+    try:
+        subprocess.run(audio_extract_command, check=True, stdout=subprocess.DEVNULL, stderr=subprocess.PIPE)
+    except subprocess.CalledProcessError:
+        print("\n   ❌ Audio extraction failed (maybe no audio?). Proceeding without audio.")
+        pass
+
+    print("\n   ✨ Step 6: Merging...")
+    if os.path.exists(temp_audio_output):
+        merge_command = [
+            'ffmpeg', '-y', '-i', temp_video_output, '-i', temp_audio_output,
+            '-c:v', 'copy', '-c:a', 'copy', final_output_video
+        ]
+    else:
+         merge_command = [
+            'ffmpeg', '-y', '-i', temp_video_output,
+            '-c:v', 'copy', final_output_video
+        ]
+
+    try:
+        subprocess.run(merge_command, check=True, stdout=subprocess.DEVNULL, stderr=subprocess.PIPE)
+        print(f"   ✅ Clip saved to {final_output_video}")
+    except subprocess.CalledProcessError as e:
+        print("\n   ❌ Final merge failed.")
+        print("   Stderr:", e.stderr.decode())
+        return False
+
+    # Clean up temp files
+    if os.path.exists(temp_video_output): os.remove(temp_video_output)
+    if os.path.exists(temp_audio_output): os.remove(temp_audio_output)
+
+    return True
