+++ openshorts/video/reframing.py
@@ -1,0 +1,42 @@
+"""Vertical reframing helpers: blurred-background 'General Shot' composite."""
+
+import cv2
+
+
+def create_general_frame(frame, output_width, output_height):
+    """
+    Creates a 'General Shot' frame:
+    - Background: Blurred zoom of original
+    - Foreground: Original video scaled to fit width, centered vertically.
+    """
+    orig_h, orig_w = frame.shape[:2]
+
+    # 1. Background (Fill Height)
+    # Crop center to aspect ratio
+    bg_scale = output_height / orig_h
+    bg_w = int(orig_w * bg_scale)
+    bg_resized = cv2.resize(frame, (bg_w, output_height))
+
+    # Crop center of background
+    start_x = (bg_w - output_width) // 2
+    if start_x < 0: start_x = 0
+    background = bg_resized[:, start_x:start_x+output_width]
+    if background.shape[1] != output_width:
+        background = cv2.resize(background, (output_width, output_height))
+
+    # Blur background
+    background = cv2.GaussianBlur(background, (51, 51), 0)
+
+    # 2. Foreground (Fit Width)
+    scale = output_width / orig_w
+    fg_h = int(orig_h * scale)
+    foreground = cv2.resize(frame, (output_width, fg_h))
+
+    # 3. Overlay
+    y_offset = (output_height - fg_h) // 2
+
+    # Clone background to avoid modifying it
+    final_frame = background.copy()
+    final_frame[y_offset:y_offset+fg_h, :] = foreground
+
+    return final_frame
