+++ openshorts/video/scene_analysis.py
@@ -1,0 +1,77 @@
+"""PySceneDetect scene boundaries + per-scene TRACK/GENERAL strategy analysis."""
+
+import cv2
+from scenedetect import open_video, SceneManager
+from scenedetect.detectors import ContentDetector
+from tqdm import tqdm
+
+from openshorts.ml.detection import detect_face_candidates
+
+
+def detect_scenes(video_path):
+    video = open_video(video_path)
+    scene_manager = SceneManager()
+    scene_manager.add_detector(ContentDetector())
+    scene_manager.detect_scenes(video=video)
+    scene_list = scene_manager.get_scene_list()
+    fps = video.frame_rate
+    return scene_list, fps
+
+
+def get_video_resolution(video_path):
+    cap = cv2.VideoCapture(video_path)
+    if not cap.isOpened():
+        raise IOError(f"Could not open video file {video_path}")
+    width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
+    height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
+    cap.release()
+    return width, height
+
+
+def analyze_scenes_strategy(video_path, scenes):
+    """
+    Analyzes each scene to determine if it should be TRACK (Single person) or GENERAL (Group/Wide).
+    Returns list of strategies corresponding to scenes.
+    """
+    cap = cv2.VideoCapture(video_path)
+    strategies = []
+
+    if not cap.isOpened():
+        return ['TRACK'] * len(scenes)
+
+    for start, end in tqdm(scenes, desc="   Analyzing Scenes"):
+        # Sample 3 frames (start, middle, end)
+        frames_to_check = [
+            start.get_frames() + 5,
+            int((start.get_frames() + end.get_frames()) / 2),
+            end.get_frames() - 5
+        ]
+
+        face_counts = []
+        for f_idx in frames_to_check:
+            cap.set(cv2.CAP_PROP_POS_FRAMES, f_idx)
+            ret, frame = cap.read()
+            if not ret: continue
+
+            # Detect faces
+            candidates = detect_face_candidates(frame)
+            face_counts.append(len(candidates))
+
+        # Decision Logic
+        if not face_counts:
+            avg_faces = 0
+        else:
+            avg_faces = sum(face_counts) / len(face_counts)
+
+        # Strategy:
+        # 0 faces -> GENERAL (Landscape/B-roll)
+        # 1 face -> TRACK
+        # > 1.2 faces -> GENERAL (Group)
+
+        if avg_faces > 1.2 or avg_faces < 0.5:
+            strategies.append('GENERAL')
+        else:
+            strategies.append('TRACK')
+
+    cap.release()
+    return strategies
