+++ openshorts/video/tracking.py
@@ -1,0 +1,212 @@
+"""SmoothedCameraman and SpeakerTracker: the heart of stabilized vertical reframing.
+
+SmoothedCameraman implements the "Heavy Tripod" pan logic — only moves when the
+subject leaves the safe zone, slow (3 px/frame) when nearby, fast (15 px/frame)
+when the distance is huge. SpeakerTracker prevents rapid speaker switching via
+hysteresis (3x score bonus for the active speaker) and a frame-based cooldown.
+"""
+
+ASPECT_RATIO = 9 / 16
+
+
+class SmoothedCameraman:
+    """
+    Handles smooth camera movement.
+    Simplified Logic: "Heavy Tripod"
+    Only moves if the subject leaves the center safe zone.
+    Moves slowly and linearly.
+    """
+    def __init__(self, output_width, output_height, video_width, video_height):
+        self.output_width = output_width
+        self.output_height = output_height
+        self.video_width = video_width
+        self.video_height = video_height
+
+        # Initial State
+        self.current_center_x = video_width / 2
+        self.target_center_x = video_width / 2
+
+        # Calculate crop dimensions once
+        self.crop_height = video_height
+        self.crop_width = int(self.crop_height * ASPECT_RATIO)
+        if self.crop_width > video_width:
+             self.crop_width = video_width
+             self.crop_height = int(self.crop_width / ASPECT_RATIO)
+
+        # Safe Zone: 20% of the video width
+        # As long as the target is within this zone relative to current center, DO NOT MOVE.
+        self.safe_zone_radius = self.crop_width * 0.25
+
+    def update_target(self, face_box):
+        """
+        Updates the target center based on detected face/person.
+        """
+        if face_box:
+            x, y, w, h = face_box
+            self.target_center_x = x + w / 2
+
+    def get_crop_box(self, force_snap=False):
+        """
+        Returns the (x1, y1, x2, y2) for the current frame.
+        """
+        if force_snap:
+            self.current_center_x = self.target_center_x
+        else:
+            diff = self.target_center_x - self.current_center_x
+
+            # SIMPLIFIED LOGIC:
+            # 1. Is the target outside the safe zone?
+            if abs(diff) > self.safe_zone_radius:
+                # 2. If yes, move towards it slowly (Linear Speed)
+                # Determine direction
+                direction = 1 if diff > 0 else -1
+
+                # Speed: 2 pixels per frame (Slow pan)
+                # If the distance is HUGE (scene change or fast movement), speed up slightly
+                if abs(diff) > self.crop_width * 0.5:
+                    speed = 15.0 # Fast re-frame
+                else:
+                    speed = 3.0  # Slow, steady pan
+
+                self.current_center_x += direction * speed
+
+                # Check if we overshot (prevent oscillation)
+                new_diff = self.target_center_x - self.current_center_x
+                if (direction == 1 and new_diff < 0) or (direction == -1 and new_diff > 0):
+                    self.current_center_x = self.target_center_x
+
+            # If inside safe zone, DO NOTHING (Stationary Camera)
+
+        # Clamp center
+        half_crop = self.crop_width / 2
+
+        if self.current_center_x - half_crop < 0:
+            self.current_center_x = half_crop
+        if self.current_center_x + half_crop > self.video_width:
+            self.current_center_x = self.video_width - half_crop
+
+        x1 = int(self.current_center_x - half_crop)
+        x2 = int(self.current_center_x + half_crop)
+
+        x1 = max(0, x1)
+        x2 = min(self.video_width, x2)
+
+        y1 = 0
+        y2 = self.video_height
+
+        return x1, y1, x2, y2
+
+
+class SpeakerTracker:
+    """
+    Tracks speakers over time to prevent rapid switching and handle temporary obstructions.
+    """
+    def __init__(self, stabilization_frames=15, cooldown_frames=30):
+        self.active_speaker_id = None
+        self.speaker_scores = {}  # {id: score}
+        self.last_seen = {}       # {id: frame_number}
+        self.locked_counter = 0   # How long we've been locked on current speaker
+
+        # Hyperparameters
+        self.stabilization_threshold = stabilization_frames # Frames needed to confirm a new speaker
+        self.switch_cooldown = cooldown_frames              # Minimum frames before switching again
+        self.last_switch_frame = -1000
+
+        # ID tracking
+        self.next_id = 0
+        self.known_faces = [] # [{'id': 0, 'center': x, 'last_frame': 123}]
+
+    def get_target(self, face_candidates, frame_number, width):
+        """
+        Decides which face to focus on.
+        face_candidates: list of {'box': [x,y,w,h], 'score': float}
+        """
+        current_candidates = []
+
+        # 1. Match faces to known IDs (simple distance tracking)
+        for face in face_candidates:
+            x, y, w, h = face['box']
+            center_x = x + w / 2
+
+            best_match_id = -1
+            min_dist = width * 0.15 # Reduced matching radius to avoid jumping in groups
+
+            # Try to match with known faces seen recently
+            for kf in self.known_faces:
+                if frame_number - kf['last_frame'] > 30: # Forgot faces older than 1s (was 2s)
+                    continue
+
+                dist = abs(center_x - kf['center'])
+                if dist < min_dist:
+                    min_dist = dist
+                    best_match_id = kf['id']
+
+            # If no match, assign new ID
+            if best_match_id == -1:
+                best_match_id = self.next_id
+                self.next_id += 1
+
+            # Update known face
+            self.known_faces = [kf for kf in self.known_faces if kf['id'] != best_match_id]
+            self.known_faces.append({'id': best_match_id, 'center': center_x, 'last_frame': frame_number})
+
+            current_candidates.append({
+                'id': best_match_id,
+                'box': face['box'],
+                'score': face['score']
+            })
+
+        # 2. Update Scores with decay
+        for pid in list(self.speaker_scores.keys()):
+             self.speaker_scores[pid] *= 0.85 # Faster decay (was 0.9)
+             if self.speaker_scores[pid] < 0.1:
+                 del self.speaker_scores[pid]
+
+        # Add new scores
+        for cand in current_candidates:
+            pid = cand['id']
+            # Score is purely based on size (proximity) now that we don't have mouth
+            raw_score = cand['score'] / (width * width * 0.05)
+            self.speaker_scores[pid] = self.speaker_scores.get(pid, 0) + raw_score
+
+        # 3. Determine Best Speaker
+        if not current_candidates:
+            # If no one found, maintain last active speaker if cooldown allows
+            # to avoid black screen or jump to 0,0
+            return None
+
+        best_candidate = None
+        max_score = -1
+
+        for cand in current_candidates:
+            pid = cand['id']
+            total_score = self.speaker_scores.get(pid, 0)
+
+            # Hysteresis: HUGE Bonus for current active speaker
+            if pid == self.active_speaker_id:
+                total_score *= 3.0 # Sticky factor
+
+            if total_score > max_score:
+                max_score = total_score
+                best_candidate = cand
+
+        # 4. Decide Switch
+        if best_candidate:
+            target_id = best_candidate['id']
+
+            if target_id == self.active_speaker_id:
+                self.locked_counter += 1
+                return best_candidate['box']
+
+            # New person
+            if frame_number - self.last_switch_frame < self.switch_cooldown:
+                old_cand = next((c for c in current_candidates if c['id'] == self.active_speaker_id), None)
+                if old_cand:
+                    return old_cand['box']
+
+            self.active_speaker_id = target_id
+            self.last_switch_frame = frame_number
+            self.locked_counter = 0
+            return best_candidate['box']
+
+        return None
